+++ src/accelerate/commands/config/config_args.py
@@ -22,7 +22,7 @@ from typing import Optional, Union
 
 import yaml
 
-from ...utils.dataclasses import ComputeEnvironment, DistributedType, SageMakerDistributedType
+from ...utils import ComputeEnvironment, DistributedType, SageMakerDistributedType
 
 
 hf_cache_home = os.path.expanduser(
