+++ src/accelerate/utils/__init__.py
@@ -3,14 +3,6 @@
 # module, but to preserve other warnings. So, don't check this module at all
 
 from .constants import MODEL_NAME, OPTIMIZER_NAME, RNG_STATE_NAME, SCALER_NAME, SCHEDULER_NAME
-from .data_helpers import (
-    find_batch_size,
-    get_data_structure,
-    honor_type,
-    initialize_tensors,
-    recursively_apply,
-    send_to_device,
-)
 from .dataclasses import (
     ComputeEnvironment,
     DeepSpeedPlugin,
@@ -19,12 +11,13 @@ from .dataclasses import (
     FullyShardedDataParallelPlugin,
     GradScalerKwargs,
     InitProcessGroupKwargs,
+    KwargsHandler,
     LoggerType,
     PrecisionType,
     RNGType,
+    SageMakerDistributedType,
     TensorInformation,
 )
-from .deepspeed import DeepSpeedEngineWrapper, DeepSpeedOptimizerWrapper
 from .imports import (
     is_apex_available,
     is_boto3_available,
@@ -37,19 +30,32 @@ from .imports import (
     is_tpu_available,
     is_wandb_available,
 )
-from .launch import PrepareForLaunch
-from .memory import find_executable_batch_size
 from .operations import (
     broadcast,
     broadcast_object_list,
     concatenate,
     convert_outputs_to_fp32,
     convert_to_fp32,
+    find_batch_size,
     gather,
     gather_object,
+    get_data_structure,
+    honor_type,
+    initialize_tensors,
+    is_tensor_information,
+    is_torch_tensor,
     pad_across_processes,
+    recursively_apply,
     reduce,
+    send_to_device,
     slice_tensors,
 )
+
+
+if is_deepspeed_available():
+    from .deepspeed import DeepSpeedEngineWrapper, DeepSpeedOptimizerWrapper
+
+from .launch import PrepareForLaunch
+from .memory import find_executable_batch_size
 from .other import extract_model_from_parallel, get_pretty_name, patch_environment, save, wait_for_everyone
 from .random import set_seed, synchronize_rng_states
